requirements.txt
django==1.10.5
djangorestframework==3.5.3
markdown
django-filter
